[flux-swf] Extract common ActivityExecutor logic into flux-common. Add generic StepInputAccessor interface.
Move all of the non-SWF-specific bits of flux-swf's ActivityExecutor to flux-common's ActivityExecutionUtil.
In particular, this will allow reuse of the pre- and post-hook execution logic for other backend implementations.

ActivityExecutionUtil and WorkflowStepUtil now take a StepInputAccessor object as the mechanism for accessing
step attributes, rather than assuming all implementations can provide a Map<String, ?> with any particular
encoding mechanism.

WorkflowStepHook inputs are still provided via Map<String, Object> since those are all generated on the fly.

When providing step inputs, Flux no longer automatically provides a mutable, empty HashMap if the requested
attribute is null or an empty string. This allows users to tell the difference between "a previous step did
not provide this attribute" and "a previous step provided an empty map".

Flux no longer guarantees that Date attributes provided by one step will be automatically converted to
Instant if a later step requests it as an Instant, and vice versa, including for the built-in attributes.
(This will still work for user-provided attributes for the time being, but that is no longer guaranteed.)
Additionally, support for the Date type as a step attribute is now documented as deprecated.

Other minor changes:
* Moved ActivityExecutionUtil from `com.danielgmyers.flux.poller` to `com.danielgmyers.flux.step.internal`
* Moved WorkflowStepUtil from `com.danielgmyers.flux.step` to `com.danielgmyers.flux.step.internal`
* Added an AttributeTypeMismatchException to be thrown if the attribute type requested by a workflow step
  does not match the actual attribute type found in the step attribute metadata. For the current SWF
  implementation this will in practice only happen for WorkflowStepHooks, since we don't actually know
  the type of the regular step attributes before we try to deserialize them as the requested type.
* flux-testutils' StepValidator no longer encodes the provided input attributes only for ActivityExecutionUtil
  to immediately decode the same attributes, by providing its own StepInputAccessor.