[flux-sfn] Implement SfnStepInputAccessor and ActivityExecutor.
SfnStepInputAccessor reads the input into a JsonNode tree and deserializes individual attributes on request.

ActivityExecutor includes the full set of input attributes in the output, because there's not a clean way to have Step Functions actually merge the input and output into the same object as part of the step definition.

Added UnsupportedAttributeTypeException for StepInputAccessors to throw as needed.

https://github.com/danielgmyers/flux-swf-client/issues/120